requirements.txt
blinker==1.8.2
flask_requests==0.0.14
pampy==0.3.0
